Grading
So, your functionality is great but the UI is kind of confusing. There's no real clues for users to know what exactly they are using. In the future the name of your app should have some hint to folks as to what exactly they are using.

Good job in the course! Have fun for the rest of the semester.